Added PortionCard and update profile editing to support Double values

+++ Snusless/Models/User.swift
@@ -11,14 +11,14 @@ import SwiftData
 @Model
 class User {
     var name: String
-    var numberOfDosor: Int
+    var numberOfDosor: Double
     var pricePerDosa: Double
     var portionsPerDosa: Int
     var savingsGoal: Int
     var startDate: Date
     var streak: Streak
     
-    init(name: String, numberOfDosor: Int, pricePerDosa: Double, portionsPerDosa: Int, savingsGoal: Int, startDate: Date,  streak: Streak = Streak()) {
+    init(name: String, numberOfDosor: Double, pricePerDosa: Double, portionsPerDosa: Int, savingsGoal: Int, startDate: Date,  streak: Streak = Streak()) {
         self.name = name
         self.numberOfDosor = numberOfDosor
         self.pricePerDosa = pricePerDosa
